ledger-on-sql: Split read queries out from write queries.
Can't run write queries from a read transaction.

CHANGELOG_BEGIN
CHANGELOG_END
